Checkpoint only if validation loss is lower

+++ h0rton/train.py
@@ -72,14 +72,11 @@ def main():
 
     # Define val data and loader
     val_data = XYData(cfg.data.val_dir, data_cfg=cfg.data)
-    val_loader = DataLoader(val_data, batch_size=cfg.optim.batch_size, shuffle=True, drop_last=True)
+    val_loader = DataLoader(val_data, batch_size=cfg.optim.batch_size, shuffle=False, drop_last=True)
     n_val = val_data.n_data
 
     # Define plotting data (subset of val data) and loader
     if cfg.log.monitor_1d_marginal_mapping:
-        plot_data_sampler = SubsetRandomSampler(np.arange(cfg.data.n_plotting))
-        plot_data_loader = DataLoader(val_data, batch_size=len(cfg.data.n_plotting), sampler=plot_data_sampler)
-        # Define plotter object
         plotter = BNNInterpreter(cfg.model.type, cfg.data.Y_dim, cfg.device)
 
     # Instantiate loss function
@@ -104,25 +101,26 @@ def main():
         epoch = 0
 
     model_path = ''
+    epoch_avg_val_loss = np.inf
+    print("Training set size: {:d}".format(n_train))
+    print("Validation set size: {:d}".format(n_val))
     progress = tqdm(range(epoch, cfg.optim.n_epochs))
     for epoch in progress:
         net.train()
         total_loss = 0.0
 
         for batch_idx, (X_, Y_) in enumerate(train_loader):
-            #X = Variable(torch.Tensor(X_)).to(cfg.device)
-            #Y = Variable(torch.Tensor(Y_)).to(cfg.device)
+            optimizer.zero_grad()
             X = X_.to(cfg.device)
             Y = Y_.to(cfg.device)
-
+            # Update weights
             pred = net(X)
             loss = loss_fn(pred, Y)
-            total_loss += loss.item()*cfg.optim.batch_size
-
-            optimizer.zero_grad()
             loss.backward()
             optimizer.step()
-        # Step every epoch
+            # For logging
+            total_loss += loss.item()*cfg.optim.batch_size
+        # Step lr_scheduler every epoch
         lr_scheduler.step()
 
         with torch.no_grad():
@@ -132,49 +130,50 @@ def main():
             for batch_idx, (X_, Y_) in enumerate(val_loader):
                 X = X_.to(cfg.device)
                 Y = Y_.to(cfg.device)
-
                 pred = net(X)
                 loss = loss_fn(pred, Y)
                 total_val_loss += loss.item()*cfg.optim.batch_size
 
             epoch_avg_train_loss = total_loss/n_train
-            epoch_avg_val_loss = total_val_loss/n_val
+            new_epoch_avg_val_loss = total_val_loss/n_val
 
             tqdm.write("Epoch [{}/{}]: TRAIN Loss: {:.4f}".format(epoch+1, cfg.optim.n_epochs, epoch_avg_train_loss))
-            tqdm.write("Epoch [{}/{}]: VALID Loss: {:.4f}".format(epoch+1, cfg.optim.n_epochs, epoch_avg_val_loss))
+            tqdm.write("Epoch [{}/{}]: VALID Loss: {:.4f}".format(epoch+1, cfg.optim.n_epochs, new_epoch_avg_val_loss))
             
             if (epoch + 1)%(cfg.log.logging_interval) == 0:
                 # Log train and val losses
                 logger.add_scalars('metrics/loss',
-                                   {'train': epoch_avg_train_loss, 'val': epoch_avg_val_loss},
+                                   {'train': epoch_avg_train_loss, 'val': new_epoch_avg_val_loss},
                                    epoch)
+                # Log alpha value
+                logger.add_histogram('val_pred/weight_gaussian2', 0.5/(torch.exp(-pred[:, -1].squeeze()) + 1.0), epoch)
 
                 # Log histograms of named parameters
-                for param_name, param in net.named_parameters():
-                    logger.add_histogram(param_name, param.clone().cpu().data.numpy(), epoch)
+                if cfg.log.monitor_weight_distributions:
+                    for param_name, param in net.named_parameters():
+                        logger.add_histogram(param_name, param.clone().cpu().data.numpy(), epoch)
 
-                # Get 1D marginal mapping plots
-                # FIXME: operation is subset of the validation loop
-                # but we want to free RAM
+                # Get performance data to monitor
+                if cfg.log.monitor_sample_images:
+                    X = X[:3].cpu().numpy()
+                    #pred = pred.cpu().numpy()
+                    logger.add_images('val_images', X, epoch, dataformats='NCHW')
+
                 if cfg.log.monitor_1d_marginal_mapping:
-                    for batch_idx, (X_, Y_) in enumerate(plot_data_loader):
-                        X_plt = X_.to(cfg.device)
-                        Y_plt = Y_.to(cfg.device)
-                        pred_plt = net(X_plt).cpu().numpy()
-
-                    plotter.set_normal_mixture_params(pred_plt)
+                    plotter.set_normal_mixture_params(pred)
                     for param_idx, param_name in enumerate(cfg.data.Y_cols):
                         tag = '1d_mapping/{:s}'.format(param_name)
-                        fig = plotter.get_1d_mapping_fig(param_name, param_idx, Y_plt[:, param_idx])
+                        fig = plotter.get_1d_mapping_fig(param_name, param_idx, Y[:, param_idx])
                         logger.add_figure(tag, fig)
 
             if (epoch + 1)%(cfg.log.checkpoint_interval) == 0:
-                os.remove(model_path) if os.path.exists(model_path) else None
-                model_path = train_utils.save_state_dict(net, optimizer, lr_scheduler, epoch_avg_train_loss, epoch_avg_val_loss, cfg.log.checkpoint_dir, cfg.model.architecture, epoch)
+                if new_epoch_avg_val_loss < epoch_avg_val_loss:
+                    os.remove(model_path) if os.path.exists(model_path) else None
+                    model_path = train_utils.save_state_dict(net, optimizer, lr_scheduler, epoch_avg_train_loss, new_epoch_avg_val_loss, cfg.log.checkpoint_dir, cfg.model.architecture, epoch)
 
     logger.close()
     os.remove(model_path) if os.path.exists(model_path) else None
-    model_path = train_utils.save_state_dict(net, optimizer, lr_scheduler, epoch_avg_train_loss, epoch_avg_val_loss, cfg.log.checkpoint_dir, cfg.model.architecture, epoch)
+    model_path = train_utils.save_state_dict(net, optimizer, lr_scheduler, epoch_avg_train_loss, new_epoch_avg_val_loss, cfg.log.checkpoint_dir, cfg.model.architecture, epoch)
     print("Saved model at {:s}".format(os.path.abspath(model_path)))
 
 if __name__ == '__main__':
